故障予測のオーバーフローバグ修正

+++ src/kumagai/smartviewer/ValueAndHourCollection.java
@@ -1,6 +1,6 @@
 package kumagai.smartviewer;
 
-import java.util.*;
+import java.util.ArrayList;
 
 import ktool.datetime.TimeSpan;
 
@@ -25,7 +25,7 @@ public class ValueAndHourCollection
 			TimeSpan realTimeDiff = get(i).datetime.diff(get(0).datetime);
 			long valueDiff = get(0).value - get(i).value;
 			int remainingHour1 = (int)(useTimeDiff * (get(i).value / valueDiff));
-			int remainingSecond2 = (int)(realTimeDiff.getTotalSecond() * (get(i).value / valueDiff));
+			long remainingSecond2 = realTimeDiff.getTotalSecond() * (get(i).value / valueDiff);
 
 			predictions.add(
 				new Prediction(
